moved analytics into data and activities into packages
