--- app/src/main/java/com/rootstrap/android/ui/adapter/Adapter.kt
@@ -1,3 +1,0 @@
-package com.rootstrap.android.ui.adapter
-
-class Adapter
