+++ app/src/main/java/com/rootstrap/android/ui/login/SignInActivityViewModel.kt
@@ -1,4 +1,4 @@
-package com.rootstrap.android.ui.activity.main
+package com.rootstrap.android.ui.login
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -6,30 +6,35 @@ import androidx.lifecycle.viewModelScope
 import com.rootstrap.android.ui.base.BaseViewModel
 import com.rootstrap.android.util.NetworkState
 import com.rootstrap.data.api.ApiException
+import com.rootstrap.data.dto.request.UserSignInRequest
+import com.rootstrap.data.dto.request.UserSignInRequestSerializer
 import com.rootstrap.data.dto.response.DataResult
 import com.rootstrap.data.managers.session.SessionManager
-import com.rootstrap.usecases.SignOut
+import com.rootstrap.usecases.SignIn
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.launch
 
-open class ProfileActivityViewModel(
-    private val signOut: SignOut,
+open class SignInActivityViewModel(
+    private val signIn: SignIn,
     private val sessionManager: SessionManager,
     uiDispatcher: CoroutineDispatcher
 ) : BaseViewModel(uiDispatcher) {
 
-    private val _state = MutableLiveData<ProfileState>()
-    val state: LiveData<ProfileState>
+    private val _state = MutableLiveData<SignInState>()
+    val state: LiveData<SignInState>
         get() = _state
 
-    fun signOut() {
+    fun signIn(user: UserSignInRequest) {
         _networkState.value = NetworkState.LOADING
         viewModelScope.launch {
-            when (val result = signOut.invoke()) {
+            when (val result = signIn.invoke(UserSignInRequestSerializer(user))) {
                 is DataResult.Success -> {
+                    result.data?.user?.let { user ->
+                        sessionManager.signIn(user)
+                    }
+
                     restoreNetworkState()
-                    _state.value = ProfileState.SIGN_OUT_SUCCESS
-                    sessionManager.signOut()
+                    _state.value = SignInState.SIGN_IN_SUCCESS
                 }
                 is DataResult.Error -> {
                     handleError(result.exception)
@@ -42,7 +47,7 @@ open class ProfileActivityViewModel(
     private fun handleError(exception: ApiException?) {
         error = exception?.message
         _networkState.value = NetworkState.ERROR
-        _state.value = ProfileState.SIGN_OUT_FAILURE
+        _state.value = SignInState.SIGN_IN_FAILURE
     }
 
     private fun restoreNetworkState() {
@@ -50,7 +55,7 @@ open class ProfileActivityViewModel(
     }
 }
 
-enum class ProfileState {
-    SIGN_OUT_FAILURE,
-    SIGN_OUT_SUCCESS
+enum class SignInState {
+    SIGN_IN_FAILURE,
+    SIGN_IN_SUCCESS
 }
