+++ app/src/main/java/com/rootstrap/android/ui/login/SignUpActivityViewModel.kt
@@ -1,4 +1,4 @@
-package com.rootstrap.android.ui.activity.main
+package com.rootstrap.android.ui.login
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -6,35 +6,37 @@ import androidx.lifecycle.viewModelScope
 import com.rootstrap.android.ui.base.BaseViewModel
 import com.rootstrap.android.util.NetworkState
 import com.rootstrap.data.api.ApiException
-import com.rootstrap.data.dto.request.UserSignInRequest
-import com.rootstrap.data.dto.request.UserSignInRequestSerializer
+import com.rootstrap.data.dto.request.UserSignUpRequest
+import com.rootstrap.data.dto.request.UserSignUpRequestSerializer
 import com.rootstrap.data.dto.response.DataResult
 import com.rootstrap.data.managers.session.SessionManager
-import com.rootstrap.usecases.SignIn
+
+import com.rootstrap.usecases.SignUp
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.launch
 
-open class SignInActivityViewModel(
-    private val signIn: SignIn,
+open class SignUpActivityViewModel(
+    private val signUp: SignUp,
     private val sessionManager: SessionManager,
     uiDispatcher: CoroutineDispatcher
 ) : BaseViewModel(uiDispatcher) {
 
-    private val _state = MutableLiveData<SignInState>()
-    val state: LiveData<SignInState>
+    private val _state = MutableLiveData<SignUpState>()
+    val state: LiveData<SignUpState>
         get() = _state
 
-    fun signIn(user: UserSignInRequest) {
+    fun signUp(userSignUpRequest: UserSignUpRequest) {
         _networkState.value = NetworkState.LOADING
         viewModelScope.launch {
-            when (val result = signIn.invoke(UserSignInRequestSerializer(user))) {
+
+            when (val result = signUp.invoke(UserSignUpRequestSerializer(userSignUpRequest))) {
                 is DataResult.Success -> {
                     result.data?.user?.let { user ->
                         sessionManager.signIn(user)
-                    }
+                    } ?: handleError(ApiException(null))
 
                     restoreNetworkState()
-                    _state.value = SignInState.SIGN_IN_SUCCESS
+                    _state.value = SignUpState.SIGN_UP_SUCCESS
                 }
                 is DataResult.Error -> {
                     handleError(result.exception)
@@ -47,7 +49,7 @@ open class SignInActivityViewModel(
     private fun handleError(exception: ApiException?) {
         error = exception?.message
         _networkState.value = NetworkState.ERROR
-        _state.value = SignInState.SIGN_IN_FAILURE
+        _state.value = SignUpState.SIGN_UP_FAILURE
     }
 
     private fun restoreNetworkState() {
@@ -55,7 +57,7 @@ open class SignInActivityViewModel(
     }
 }
 
-enum class SignInState {
-    SIGN_IN_FAILURE,
-    SIGN_IN_SUCCESS
+enum class SignUpState {
+    SIGN_UP_FAILURE,
+    SIGN_UP_SUCCESS
 }
