+++ app/src/main/java/com/rootstrap/android/ui/profile/ProfileActivityViewModel.kt
@@ -1,4 +1,4 @@
-package com.rootstrap.android.ui.activity.main
+package com.rootstrap.android.ui.profile
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -6,37 +6,30 @@ import androidx.lifecycle.viewModelScope
 import com.rootstrap.android.ui.base.BaseViewModel
 import com.rootstrap.android.util.NetworkState
 import com.rootstrap.data.api.ApiException
-import com.rootstrap.data.dto.request.UserSignUpRequest
-import com.rootstrap.data.dto.request.UserSignUpRequestSerializer
 import com.rootstrap.data.dto.response.DataResult
 import com.rootstrap.data.managers.session.SessionManager
-
-import com.rootstrap.usecases.SignUp
+import com.rootstrap.usecases.SignOut
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.launch
 
-open class SignUpActivityViewModel(
-    private val signUp: SignUp,
+open class ProfileActivityViewModel(
+    private val signOut: SignOut,
     private val sessionManager: SessionManager,
     uiDispatcher: CoroutineDispatcher
 ) : BaseViewModel(uiDispatcher) {
 
-    private val _state = MutableLiveData<SignUpState>()
-    val state: LiveData<SignUpState>
+    private val _state = MutableLiveData<ProfileState>()
+    val state: LiveData<ProfileState>
         get() = _state
 
-    fun signUp(userSignUpRequest: UserSignUpRequest) {
+    fun signOut() {
         _networkState.value = NetworkState.LOADING
         viewModelScope.launch {
-
-            when (val result = signUp.invoke(UserSignUpRequestSerializer(userSignUpRequest))) {
+            when (val result = signOut.invoke()) {
                 is DataResult.Success -> {
-                    result.data?.user?.let { user ->
-                        sessionManager.signIn(user)
-                    } ?: handleError(ApiException(null))
-
                     restoreNetworkState()
-                    _state.value = SignUpState.SIGN_UP_SUCCESS
+                    _state.value = ProfileState.SIGN_OUT_SUCCESS
+                    sessionManager.signOut()
                 }
                 is DataResult.Error -> {
                     handleError(result.exception)
@@ -49,7 +42,7 @@ open class SignUpActivityViewModel(
     private fun handleError(exception: ApiException?) {
         error = exception?.message
         _networkState.value = NetworkState.ERROR
-        _state.value = SignUpState.SIGN_UP_FAILURE
+        _state.value = ProfileState.SIGN_OUT_FAILURE
     }
 
     private fun restoreNetworkState() {
@@ -57,7 +50,7 @@ open class SignUpActivityViewModel(
     }
 }
 
-enum class SignUpState {
-    SIGN_UP_FAILURE,
-    SIGN_UP_SUCCESS
+enum class ProfileState {
+    SIGN_OUT_FAILURE,
+    SIGN_OUT_SUCCESS
 }
