+++ app/src/main/res/layout/activity_sign_up.xml
@@ -9,7 +9,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_margin="@dimen/sign_up_margin_horizontal"
-        tools:context=".ui.activity.main.SignUpActivity">
+        tools:context=".ui.login.SignUpActivity">
 
         <com.google.android.material.textfield.TextInputLayout
             android:id="@+id/first_name_text_input_layout"
